Aggiunto stato a piano, aggiornate query per azioni ed iniziata implementazione home piano
