+++ strt/serapide_client/src/components/StatoProgress.js
@@ -8,20 +8,23 @@
 import React from 'react'
 import { UncontrolledTooltip } from 'reactstrap'
 import shortid from "shortid"
-const getActive = (stato, currentStato  = "" ) => stato === currentStato.toLowerCase() ? "active" : ''
+import {fasi} from "../utils"
+import classNames from "classnames"
 
-export default ({stato: {nome = "Unknown", codice, descrizione}}) => {
+export default ({stato: {nome = "Unknown", codice}, className= "stato-progress", legend= false}) => {
     const id =  `_${shortid.generate()}`
+    const activeFase = fasi[fasi.indexOf(nome.toLocaleLowerCase()) + 1]
     return (
-        <span className="stato-progress">
-            <i id={id} className={`material-icons ${nome.toLowerCase()}`}>room</i>
-            <ul className={`_${nome.toLowerCase()}`}>
-                <li></li>
-                <li id="_avvio" className={getActive("avvio", nome)}></li>
-                <li id="_adozione" className={getActive("adozione", nome)}></li>
-                <li id="_approvazione" className={getActive("approvazione", nome)}></li>
-                <li id="_pubblicazione" className={getActive("pubblicazione", nome)}></li>
-                <UncontrolledTooltip placement="top" target={id} ><span className="text-capitalize">{nome.toLowerCase()}</span></UncontrolledTooltip>
+        <span className={className}>
+            <i id={id} className={`material-icons ${activeFase}`}>room</i>
+            <ul className={`_${activeFase} _icons`}>
+                {fasi.map((fase, idx) => (<li key={`_${fase}`} id={`_${fase}`} className={classNames({active: activeFase === fase})}/>))}
+                {!legend && <UncontrolledTooltip placement="top" target={id} ><span className="text-capitalize">{nome.toLowerCase()}</span></UncontrolledTooltip>}
             </ul>
+            {legend && (
+                <React.Fragment>
+                <span className={classNames("current-fase", "text-capitalize", activeFase)}>{activeFase}</span>
+                {fasi.map((fase) => (<UncontrolledTooltip key={`_${fase}`} placement="bottom" target={`_${fase}`} ><span className="text-capitalize">{fase}</span></UncontrolledTooltip>))}
+                </React.Fragment>)}
         </span>)
     }
