+++ strt/serapide_client/src/pages/Piano.js
@@ -15,7 +15,8 @@ import {getEnteLabel, getPianoLabel} from "../utils"
 import {GET_PIANI} from "../queries"
 import Injector from '../components/Injector'
 import SideBar from '../components/SideBarMenu'
-const HOME = Anagrafica
+import StatoProgress from '../components/StatoProgress'
+import Home from "./Home"
 
 const getActive = (url = "", pathname = "") => {
     return pathname.replace(url, "")
@@ -46,12 +47,17 @@ export default ({match: {url, path, params: {code} = {}} = {},location: {pathnam
                 </Injector>
                 <div className="serapide-content pt-5 pb-5 pX-md px-1 serapide-top-offset position-relative overflow-x-scroll">
                     <div className="d-flex flex-column ">
-                        <h4 className="text-uppercase">{getEnteLabel(piano.ente)}</h4>  
-                        <h2 className="mb-0 text-capitalize">{`${getPianoLabel(piano.tipo)} ${code}`}</h2>  
-                        <h5 className="text-capitalize">{getPianoLabel.descrizione}</h5>  
-                        
+                        <div className="d-flex justify-content-between align-items-center ">
+                            <div>
+                                <h4 className="text-uppercase">{getEnteLabel(piano.ente)}</h4>  
+                                <h2 className="mb-0 text-capitalize">{`${getPianoLabel(piano.tipo)} ${code}`}</h2>  
+                                <h5 className="text-capitalize">{getPianoLabel.descrizione}</h5>  
+                            </div>
+                            <StatoProgress className="stato-progress-xxl" stato={piano.fase} legend></StatoProgress>
+                        </div>
                         <Route  path={`${path}/anagrafica`} component={Anagrafica}/>
-                        <Route  path={`${path}`} component={HOME}/>
+                        <Route  path={`${path}`} component={Home}/>
+                        
                         
                             
                     </div>
