+++ strt/serapide_client/src/utils.js
@@ -5,3 +5,4 @@ export const getEnteLabel = ({name = "", code, type: {tipoente = ""} ={}} = {}) 
 export const getEnteLabelID = ({name = "", code, type: {tipoente = ""} ={}} = {}) => `ID ${tipoente} ${code}`
 export const getPianoLabel = (tipo = "") => (tipo === "VARIANTE" ? tipo.toLowerCase() : `piano ${tipo.toLowerCase()}`)
 export const formatDate = (date) => format(parseISO(date) , "iiii do MMMM yyyy", {locale: it})
+export const fasi = [ "draft", "anagrafica", "avvio", "adozione", "approvazione", "pubblicazione"]
